Move Rex::Job into its own file
